requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.7
django-appconf==1.0.1
django-bootstrap3==5.2.0
django-compressor==1.4
django-toolbelt==0.0.1
djangorestframework==3.1.1
drf-extensions==0.2.7
gunicorn==19.3.0
psycopg2==2.6
six==1.9.0
static3==0.5.1
whitenoise==1.0.6
